requirements.txt
# requirements.txt

python-dotenv
requests
sendgrid==6.0.5

Flask